+++ Core.lua
@@ -18,7 +18,7 @@ local ADDON_NAME, CN = ...
 _G.CompletionNavigator = CN
 
 CN.name        = ADDON_NAME
-CN.version     = "0.67.1"
+CN.version     = "0.67.2"
 CN.dbVersion   = 21
 
 -- Where the addon's own textures live. Referenced by the .toc IconTexture